+++ Libplanet.Tests/Net/Protocols/RoutingTableTest.cs
@@ -16,6 +16,10 @@ namespace Libplanet.Tests.Net.Protocols
         private const int BucketSize = 16;
         private const int TableSize = Address.Size * sizeof(byte) * 8;
 
+        private static readonly PrivateKey VersionSigner = new PrivateKey();
+        private static readonly AppProtocolVersion AppProtocolVer =
+            AppProtocolVersion.Sign(VersionSigner, 1);
+
         [Fact]
         public void BucketTest()
         {
@@ -23,23 +27,23 @@ namespace Libplanet.Tests.Net.Protocols
             var peer1 = new BoundPeer(
                 new PrivateKey().PublicKey,
                 new DnsEndPoint("0.0.0.0", 1234),
-                0);
+                AppProtocolVer);
             var peer2 = new BoundPeer(
                 new PrivateKey().PublicKey,
                 new DnsEndPoint("0.0.0.0", 1234),
-                0);
+                AppProtocolVer);
             var peer3 = new BoundPeer(
                 new PrivateKey().PublicKey,
                 new DnsEndPoint("0.0.0.0", 1234),
-                0);
+                AppProtocolVer);
             var peer4 = new BoundPeer(
                 new PrivateKey().PublicKey,
                 new DnsEndPoint("0.0.0.0", 1234),
-                0);
+                AppProtocolVer);
             var peer5 = new BoundPeer(
                 new PrivateKey().PublicKey,
                 new DnsEndPoint("0.0.0.0", 1234),
-                0);
+                AppProtocolVer);
 
             Assert.Empty(bucket.Peers);
             Assert.True(bucket.IsEmpty());
@@ -100,7 +104,7 @@ namespace Libplanet.Tests.Net.Protocols
         {
             var pubKey = new PrivateKey().PublicKey;
             RoutingTable table = CreateTable(pubKey.ToAddress());
-            var peer = new BoundPeer(pubKey, new DnsEndPoint("0.0.0.0", 1234), 0);
+            var peer = new BoundPeer(pubKey, new DnsEndPoint("0.0.0.0", 1234), AppProtocolVer);
             Assert.Throws<ArgumentException>(() => table.AddPeer(peer));
         }
 
@@ -125,9 +129,9 @@ namespace Libplanet.Tests.Net.Protocols
                 2,
                 new System.Random(),
                 Logger.None);
-            var peer1 = new BoundPeer(pubKey1, new DnsEndPoint("0.0.0.0", 1234), 0);
-            var peer2 = new BoundPeer(pubKey2, new DnsEndPoint("0.0.0.0", 1234), 0);
-            var peer3 = new BoundPeer(pubKey3, new DnsEndPoint("0.0.0.0", 1234), 0);
+            var peer1 = new BoundPeer(pubKey1, new DnsEndPoint("0.0.0.0", 1234), AppProtocolVer);
+            var peer2 = new BoundPeer(pubKey2, new DnsEndPoint("0.0.0.0", 1234), AppProtocolVer);
+            var peer3 = new BoundPeer(pubKey3, new DnsEndPoint("0.0.0.0", 1234), AppProtocolVer);
             table.AddPeer(peer1);
             table.AddPeer(peer2);
             table.AddPeer(peer3);
@@ -150,8 +154,8 @@ namespace Libplanet.Tests.Net.Protocols
                 2,
                 new System.Random(),
                 Logger.None);
-            var peer1 = new BoundPeer(pubKey1, new DnsEndPoint("0.0.0.0", 1234), 0);
-            var peer2 = new BoundPeer(pubKey2, new DnsEndPoint("0.0.0.0", 1234), 0);
+            var peer1 = new BoundPeer(pubKey1, new DnsEndPoint("0.0.0.0", 1234), AppProtocolVer);
+            var peer2 = new BoundPeer(pubKey2, new DnsEndPoint("0.0.0.0", 1234), AppProtocolVer);
 
             Assert.Throws<ArgumentException>(() => table.RemovePeer(peer1));
             Assert.Throws<ArgumentNullException>(() => table.RemovePeer(null));